feat(plugins): replace Generator plugins with Template plugins
A Template plugin can have multiple templates and each template has a specific file extension.
Which template to pick for new migrations are currently controlled by using the "extension" option,
and later an interactive prompt will show up in case of collisions (i.e. when multiple plugins have templates for the same extension).
Another feature in this change is that Loader plugins can chose to return `undefined` to tell Emigrate to skip to the next plugin,
which opens up the possibility to have multiple Loader plugins registered that can load the same file extension.
It is up to each Loader plugin to figure out if they can load a certain file or not,
and they will only be called for their registered file extensions.