feat(new): add "prefix" and "joiner" options to the new command
With the prefix option it's possible to customize the generation of the filename prefix for new migration files.
There are several built-in prefix generators and custom generators can be configured in the configuration file.

BREAKING CHANGE: the "-x" CLI option is now an alias for "--prefix" whereas "--extension" now has the alias "-e" instead.