refactor: drop the "directory" property from the ReporterInitParameters
This is to make Emigrate more flexible when it comes to migration locations.
Future versions of Emigrate will probably support doing migrations that are not file based,
and in such case a "directory" makes no sense.

BREAKING CHANGE: the "directory" option is no longer passed as a parameter to the reporter. The information is still available for each migration if needed, but that will likely change in a future version as well