feat(plugins): do a complete rewrite of the plugin system (WIP)
Instead of having different kinds of plugins there's now a plugin inspired by Astro's integrations that's based on hooks.
Each plugin can decide which hooks to implement and depending on what the plugin decides to do in each hook its role varies.
For each migration command a context is built and when it's run it's going through different phases,
where the first phase is loading config and plugins, which themselves can modify the config (to for instance automatically add additional plugins).
The next phase is for collecting and loading migrations where each plugin can register migrations to run (the collect step) and how to do it (the load step).
Thanks to this it's now possible for plugins to run their own migrations as part of the migration,
for instance if a plugin needs/wants to change the data model for its history table.
With this Emigrate is no longer relying on migrations being files either,
a migration can be as simple as an ordinary function.
To ease registering functions as migrations (and to for instance run Emigrate without a CLI) there's now a built-in "inline" plugin that can do that.
With this the "directory" option is no longer required and to get the old behaviour the built-in "directory" plugin should be used instead.
As plugins can both read and modify the configuration I might change the config interface so that the "directory" plugin is automatically loaded and used if a "directory" field is set in the config, so it's similar to the old behaviour.
In this change no existing plugins have been changed yet though and the whole feature is a work in progress.
So there's no actual breaking change just yet, as this is built in parallel with the existing functionality.